+++ jax/_src/pjit.py
@@ -367,7 +367,7 @@ def _parse_jit_arguments(fun: Callable, *, in_shardings: Any,
         'backend and device argument on jit is deprecated. You can use'
         ' `jax.device_put(..., jax.local_devices(backend="cpu")[0])` on the'
         ' inputs to the jitted function to get the same behavior.',
-        DeprecationWarning,
+        category=DeprecationWarning, stacklevel=2
     )
     if device is not None and backend is not None:
       raise ValueError("can't specify both a device and a backend for jit, "