+++ tests/multiprocess/multihost_utils_test.py
@@ -171,6 +171,8 @@ class MultiHostUtilsTest(jt_multiprocess.MultiProcessTest):
       else:
         multihost_utils.sync_global_devices('test message2')
 
+  @jtu.ignore_warning(category=DeprecationWarning,
+                      message='`with mesh:` context manager')
   def test_sync_global_devices_mesh_context_manager(self):
     global_mesh = jtu.create_mesh((2, 2), ('x', 'y'), iota_order=True)
     with global_mesh: