+++ tests/pjit_test.py
@@ -95,8 +95,9 @@ def check_1d_2d_mesh(f, set_mesh):
     ))(jtu.with_mesh_from_kwargs(f) if set_mesh else f)
 
 
-# TODO(skye): make the buffer donation utils part of JaxTestCase
 @jtu.pytest_mark_if_available('multiaccelerator')
+@jtu.ignore_warning(category=DeprecationWarning,
+                    message='`with mesh:` context manager')
 class PJitTest(jtu.BufferDonationTestCase):
 
   @jtu.with_mesh([('x', 1)])
@@ -1491,6 +1492,8 @@ class AutoShardingPjitTest(jtu.JaxTestCase):
 
 
 @jtu.pytest_mark_if_available('multiaccelerator')
+@jtu.ignore_warning(category=DeprecationWarning,
+                    message='`with mesh:` context manager')
 class ArrayPjitTest(jtu.JaxTestCase):
 
   @parameterized.named_parameters(
@@ -9823,6 +9826,8 @@ class ShardingInTypesTest(jtu.JaxTestCase):
 
 
 @jtu.pytest_mark_if_available('multiaccelerator')
+@jtu.ignore_warning(category=DeprecationWarning,
+                    message='`with mesh:` context manager')
 class PJitErrorTest(jtu.JaxTestCase):
 
   @check_1d_2d_mesh(set_mesh=True)