+++ CHANGELOG.md
@@ -19,6 +19,8 @@ When releasing, please add the new-release-boilerplate to docs/pallas/CHANGELOG.
 * Deprecations
   * `jax.lax.pvary` has been deprecated.
     Please use `jax.lax.pcast(..., to='varying')` as the replacement.
+  * `with mesh:` context manager has been deprecated.
+    Please use `with jax.set_mesh(mesh):` instead.
 
 * Changes:
   * jax's `Tracer` no longer inherits from `jax.Array` at runtime. However,